refactor: Migrate verify-email page within auth
