+++ src/app/auth/error/page.tsx
@@ -5,33 +5,33 @@ import Link from 'next/link'
 import { Suspense } from 'react'
 
 function AuthErrorComponent() {
-    const searchParams = useSearchParams()
-    const errorMessage = searchParams?.get('message') || 'An unknown error occurred'
+  const searchParams = useSearchParams()
+  const errorMessage = searchParams?.get('message') || 'An unknown error occurred'
 
-    return (
-        <div className="min-h-screen flex items-center justify-center bg-gray-50">
-            <div className="max-w-md w-full p-6 bg-white rounded-lg shadow-lg">
-                <div className="text-center">
-                    <h1 className="text-red-600 text-4xl font-bold mb-4">Authentication Error</h1>
-                    <div className="bg-red-50 border border-red-200 rounded-md p-4 mb-4">
-                        <p className="text-red-700">{decodeURIComponent(errorMessage)}</p>
-                    </div>
-                    <Link
-                        href="/api/auth/login"
-                        className="inline-block bg-blue-600 hover:bg-blue-700 text-white font-medium py-2 px-4 rounded-md transition-colors"
-                    >
-                        Return to Sign In
-                    </Link>
-                </div>
-            </div>
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50">
+      <div className="max-w-md w-full p-6 bg-white rounded-lg shadow-lg">
+        <div className="text-center">
+          <h1 className="text-red-600 text-4xl font-bold mb-4">Authentication Error</h1>
+          <div className="bg-red-50 border border-red-200 rounded-md p-4 mb-4">
+            <p className="text-red-700">{decodeURIComponent(errorMessage)}</p>
+          </div>
+          <Link
+            href="/api/auth/login"
+            className="inline-block bg-blue-600 hover:bg-blue-700 text-white font-medium py-2 px-4 rounded-md transition-colors"
+          >
+            Return to Sign In
+          </Link>
         </div>
-    )
+      </div>
+    </div>
+  )
 }
 
 export default function AuthError() {
-    return (
-        <Suspense fallback={<div>Loading...</div>}>
-            <AuthErrorComponent />
-        </Suspense>
-    )
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <AuthErrorComponent />
+    </Suspense>
+  )
 }
